icode: only force uploading files for 1st submission
Before, we'd do a first upload for all the icode activities when the VM
was ready. It seems overkill especially if the user won't use all the
activities right away, and it might cause some latency upon loading the
book.

Instead, let's force the upload at the 1st submission. When the VM isn't
ready, the submit button is disabled.

Signed-off-by: Joël Porquet-Lupine <joel@porquet.org>