npm: install JS linter/formatter
Use a linter/formatter to make our JS source more consistent.

Note that prettier is run as a eslint plugin, this way we get real-time
formatting feedback in the code editor, rather than formatting file as a
discrete step.

Note also that prettier is saved as an exact version, as per the install
instructions.

Signed-off-by: Joël Porquet-Lupine <joel@porquet.org>